# cedar-artifact-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Better Metadata - Better Science\n\nInternal service: business routes require both X-CEDAR-Artifact-Service-Key and the\nend-user Authorization header. Public callers use resource; a user API key alone is insufficient.\n\nThe artifact server stores templates, template elements, template fields, and template\ninstances, and validates every one of them against the CEDAR model before storing it.\n\nArtifacts use strong ETags for optimistic concurrency. Read the artifact and send the returned\nETag in `If-Match` when replacing or deleting it. A missing required precondition returns 428;\na stale or malformed validator returns 412. `If-Match: *` accepts whichever revision currently\nexists and therefore deliberately gives up lost-update protection. Representation suffixes such\nas `\"7-yaml\"` and `\"7-yaml-compact-v2\"` identify different bytes of the same stored revision and\nare valid write preconditions.\n\nArtifacts are read and written as JSON or YAML, selected by the Accept and Content-Type\nheaders. The compact YAML form is a lossy read-time convenience: it can be read but not stored.\n",
    "title" : "CEDAR Artifact Server API",
    "version" : "2.9.11-SNAPSHOT"
  },
  "paths" : {
...
    "/template-fields" : {
      "get" : {
        "description" : "List template fields, one page at a time. The response carries the size of the whole collection and paging links, not just the returned page.",
        "operationId" : "findAllTemplateFields",
        "parameters" : [ {
          "description" : "Maximum number of artifacts to return.",
          "in" : "query",
          "name" : "limit",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "description" : "Number of artifacts to skip before the first one returned.",
          "in" : "query",
          "name" : "offset",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "description" : "Return only the summary fields configured for this artifact type.",
          "in" : "query",
          "name" : "summary",
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "description" : "Comma-separated list of fields to return, in place of the default selection. Ignored when `summary` is set.",
          "in" : "query",
          "name" : "field_names",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/SchemaArtifactDocument"
                  }
                }
              }
            },
            "description" : "A page of template fields",
            "headers" : {
              "Link" : {
                "description" : "Paging links for the first, previous, next, and last pages, where those exist.",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              },
              "Total-Count" : {
                "description" : "Total number of artifacts of this type, which is the whole collection rather than the returned page.",
                "schema" : {
                  "type" : "integer"
                },
                "style" : "simple"
              }
            }
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "A paging parameter is out of range, or field_names is combined with summary=true"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ],
          "artifact_service" : [ ]
        } ],
        "summary" : "List template fields",
        "tags" : [ "Template fields" ]
      },
      "post" : {
        "description" : "Create a template field. The body may be JSON or YAML, selected by the Content-Type header. A YAML body must be the full or minimal form. The artifact is validated against the CEDAR model before it is stored; an invalid one is refused rather than stored. The server mints the identifier, so the body must carry none, and must carry a name.",
        "operationId" : "createTemplateField",
        "parameters" : [ {
          "description" : "Not supported on writes, and rejected with 400. A write response always renders the full form, and the compact form can not be stored.",
          "in" : "query",
          "name" : "compact",
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SchemaArtifactDocument"
              }
            },
            "application/x-yaml" : {
              "schema" : {
                "$ref" : "#/components/schemas/SchemaArtifactDocument"
              }
            },
            "application/yaml" : {
              "schema" : {
                "$ref" : "#/components/schemas/SchemaArtifactDocument"
              }
            }
          },
          "description" : "The template field to store.",
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SchemaArtifactDocument"
                }
              },
              "application/x-yaml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SchemaArtifactDocument"
                }
              },
              "application/yaml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SchemaArtifactDocument"
                }
              }
            },
            "description" : "The stored template field",
            "headers" : {
              "CEDAR-Validation-Status" : {
                "description" : "Whether the artifact passed validation against the CEDAR model. Every write is validated; an invalid artifact is refused rather than stored.",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              },
              "ETag" : {
                "description" : "Strong validator for the returned representation and its stored revision. A representation suffix such as \"7-yaml\" identifies different bytes of the same stored revision and is a valid write precondition.",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              },
              "Location" : {
                "description" : "URL of the created artifact.",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              },
              "application/x-yaml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              },
              "application/yaml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The body is empty, carries an identifier, has no name, or failed validation"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              },
              "application/x-yaml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              },
              "application/yaml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              },
              "application/x-yaml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              },
              "application/yaml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Forbidden"
          },
          "406" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              },
              "application/x-yaml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              },
              "application/yaml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The Accept header names no representation this artifact can be served in."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              },
              "application/x-yaml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              },
              "application/yaml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ],
          "artifact_service" : [ ]
        } ],
        "summary" : "Create a template field",
        "tags" : [ "Template fields" ]
      }
    },
...
  },
  "components" : {
...
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      },
      "artifact_service" : {
        "description" : "Private installation credential for trusted artifact callers; required alongside user authorization.",
        "in" : "header",
        "name" : "X-CEDAR-Artifact-Service-Key",
        "type" : "apiKey"
      }
    }
  }
}